% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/africontinent.r
\docType{data}
\name{africontinent}
\alias{africontinent}
\alias{sfpolygon}
\title{African country boundaries}
\format{
Formal class 'sf';
\itemize{
\item{name} {character vector continent name}
\item{geometry} {sfc_MULTIPOLYGON}
}

Geographical coordinates WGS84 datum (CRS EPSG 4326)
}
\source{
\url{https://www.naturalearthdata.com/downloads/110m-cultural-vectors/}
}
\usage{
africontinent
}
\description{
a \code{sf} object containing low resolution African continent outline
See data-raw/afrilearndata-creation.R for how the data object is created.
}
\examples{
if (requireNamespace("sf", quietly = TRUE)) {
  library(sf)
  data(africontinent)
  # or
  filename <- system.file("extdata","africontinent.shp", package="afrilearndata", mustWork=TRUE)
  africontinent <- sf::read_sf(filename)

  plot(sf::st_geometry(africontinent))
}
}
\seealso{
https://cran.r-project.org/web/packages/rnaturalearth/
}
\keyword{datasets}
\keyword{sf}
