% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastafriwpop.r
\docType{data}
\name{afripop2000}
\alias{afripop2000}
\alias{afripop00}
\title{modelled population density 2000 per square km from WorldPop aggregated to mean per 20km squares}
\format{
Formal class 'raster';

Geographical coordinates WGS84 datum (CRS EPSG 4326)
}
\source{
\url{https://www.worldpop.org/geodata/summary?id=24757}
}
\usage{
afripop2000
}
\description{
a \code{raster} object modelled population density 2000 per square km from WorldPop aggregated to mean per 20km squares
}
\examples{
if (requireNamespace("raster", quietly = TRUE)) {
  library(raster)
  data(afripop2000)
  # or
  filename <- system.file("extdata/afripop2000.tif", package="afrilearndata")
  afripop2000 <- raster::raster(filename)

  plot(afripop2000)
}
}
\seealso{
https://www.worldpop.org
WorldPop datasets are licensed under Creative Commons Attribution 4.0 International (CC BY 4.0) https://creativecommons.org/licenses/by/4.0/
}
\keyword{datasets}
\keyword{sf}
