% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afripop.r
\docType{data}
\name{afripop2020}
\alias{afripop2020}
\alias{afripop20}
\title{modelled population density 2020 per square km from WorldPop aggregated to mean per 20km squares}
\format{
Formal class 'raster';

Geographical coordinates WGS84 datum (CRS EPSG 4326)
}
\source{
\url{https://www.worldpop.org/geodata/summary?id=24777}
}
\usage{
afripop2020
}
\description{
a \code{raster} object modelled population density 2020 per square km from WorldPop aggregated to mean per 20km squares
}
\examples{
if (requireNamespace("raster", quietly = TRUE)) {
  library(raster)
  data(afripop2020)
  # or
  filename <- system.file("extdata","afripop2020.tif", package="afrilearndata", mustWork=TRUE)
  afripop2020 <- raster::raster(filename)

  plot(afripop2020)
}
}
\seealso{
https://www.worldpop.org
WorldPop datasets are licensed under Creative Commons Attribution 4.0 International (CC BY 4.0) https://creativecommons.org/licenses/by/4.0/
}
\keyword{datasets}
\keyword{sf}
