% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastafriwpop.r
\docType{data}
\name{rastafriwpop2000}
\alias{rastafriwpop2000}
\alias{rastafriwpop00}
\title{modelled population density 2000 per square km from WorldPop aggregated to mean per 20km squares}
\format{
Formal class 'raster';

Geographical coordinates WGS84 datum (CRS EPSG 4326)
}
\source{
\url{https://www.worldpop.org/geodata/summary?id=24757}
}
\usage{
rastafriwpop2000
}
\description{
a \code{raster} object modelled population density 2000 per square km from WorldPop aggregated to mean per 20km squares
}
\examples{
if (requireNamespace("raster", quietly = TRUE)) {
  library(raster)
  data(rastafriwpop2000)
  # or
  filename <- system.file("extdata/rastafriwpop2000.tif", package="afrilearndata")
  rastafriwpop2000 <- raster::raster(filename)

  plot(rastafriwpop2000)
}
}
\seealso{
https://www.worldpop.org
}
\keyword{datasets}
\keyword{sf}
