% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastafriwpop.r
\docType{data}
\name{rastafriwpop2020}
\alias{rastafriwpop2020}
\alias{rastafriwpop20}
\title{modelled population density 2020 per square km from WorldPop aggregated to mean per 20km squares}
\format{
Formal class 'raster';

Geographical coordinates WGS84 datum (CRS EPSG 4326)
}
\source{
\url{https://www.worldpop.org/geodata/summary?id=24777}
}
\usage{
rastafriwpop2020
}
\description{
a \code{raster} object modelled population density 2020 per square km from WorldPop aggregated to mean per 20km squares
}
\examples{
if (requireNamespace("raster", quietly = TRUE)) {
  library(raster)
  data(rastafriwpop2020)
  # or
  filename <- system.file("extdata/rastafriwpop2020.tif", package="afrilearndata")
  rastafriwpop2020 <- raster::raster(filename)

  plot(rastafriwpop2020)
}
}
\seealso{
https://www.worldpop.org
}
\keyword{datasets}
\keyword{sf}
