% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfafrihway.r
\docType{data}
\name{sfafrihway}
\alias{sfafrihway}
\alias{sflines}
\title{African trans-continental highway network lines}
\format{
Formal class 'sf' \link{package "sf"};
\itemize{
\item{name_long} {character vector of country names}
\item{geom} {sfc_MULTIPOLYGON}
}
The object is in geographical coordinates using the WGS84 datum.
}
\source{
\url{https://www.google.com/maps/d/u/0/viewer?msa=0&mid=1nEU2oBFzSxabx3Z14nTyZP3KSzY&ll=1.9249940151081273\%2C12.874260000000021&z=3}
}
\usage{
sfafrihway
}
\description{
The object loaded is a \code{sf} object containing simplified lines of planned transcontinental highway
}
\examples{
if (requireNamespace("sf", quietly = TRUE)) {
  library(sf)
  data(sfafrihway)
  # or
  sfafrihway <- sf::read_sf("inst/extdata/Trans-African Highway Network.kml")

  plot(sf::st_geometry(sfafrihway))
}
}
\seealso{
https://en.wikipedia.org/wiki/Trans-African_Highway_network
}
\keyword{datasets}
\keyword{sf}
